Debian Wheezy and Jessie are working

+++ scripts/puppet.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 
-# Prepare puppetlabs repo
-wget http://apt.puppetlabs.com/puppetlabs-release-wheezy.deb
-dpkg -i puppetlabs-release-wheezy.deb
-apt-get -y update
+if [ $(facter lsbdistcodename) = 'wheezy' ] ; then
+    # Prepare puppetlabs repo
+    wget http://apt.puppetlabs.com/puppetlabs-release-wheezy.deb
+    dpkg -i puppetlabs-release-wheezy.deb
+    rm -f puppetlabs-release-wheezy.deb
+    apt-get -y update
+fi
 
 # Install puppet
 apt-get -y install puppet facter
-
-# Cleanup
-rm -f puppetlabs-release-wheezy.deb
